# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EquipmentStatus = typeof EquipmentStatus[keyof typeof EquipmentStatus];


export const EquipmentStatus = {
  running: 'running',
  warning: 'warning',
  fault: 'fault',
  offline: 'offline',
  maintenance: 'maintenance',
} as const;

/**
 * @nullable
 */
export type MetricTrend = typeof MetricTrend[keyof typeof MetricTrend] | null;


export const MetricTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface Metric {
  key: string;
  label: string;
  value: number;
  unit: string;
  timestamp: string;
  /** @nullable */
  trend?: MetricTrend;
}

export interface Equipment {
  id: string;
  /** Chinese name */
  name: string;
  /** English name */
  nameEn: string;
  category: string;
  status: EquipmentStatus;
  location: string;
  locationEn?: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  lastMaintenance: string;
  /** @nullable */
  predictedFailureDate?: string | null;
  metrics?: Metric[];
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  critical: 'critical',
  warning: 'warning',
  info: 'info',
} as const;

export interface Alert {
  id: string;
  equipmentId: string;
  equipmentName: string;
  severity: AlertSeverity;
  /** Chinese message */
  message: string;
  /** English message */
  messageEn: string;
  timestamp: string;
  acknowledged: boolean;
}

export interface EquipmentCategory {
  id: string;
  name: string;
  nameEn: string;
  icon: string;
  count: number;
  alertCount: number;
}

export interface DashboardSummary {
  totalEquipment: number;
  runningCount: number;
  warningCount: number;
  faultCount: number;
  offlineCount: number;
  maintenanceCount: number;
  activeAlerts: number;
  criticalAlerts: number;
  /** Overall system health 0-100 */
  healthScore: number;
}

export interface StatusBreakdown {
  running: number;
  warning: number;
  fault: number;
  offline: number;
  maintenance: number;
}

export interface EnvironmentMetrics {
  temperature: number;
  humidity: number;
  vibration: number;
  power: number;
  timestamp: string;
}

export type ListEquipmentParams = {
category?: string;
status?: string;
search?: string;
};

export type ListAlertsParams = {
acknowledged?: boolean;
severity?: string;
};