# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface Alert {
  id: string;
  equipmentId: string;
  equipmentName: string;
  severity: AlertSeverity;
  /** Chinese message */
  message: string;
  /** English message */
  messageEn: string;
  timestamp: string;
  acknowledged: boolean;
}