# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalEquipment: number;
  runningCount: number;
  warningCount: number;
  faultCount: number;
  offlineCount: number;
  maintenanceCount: number;
  activeAlerts: number;
  criticalAlerts: number;
  /** Overall system health 0-100 */
  healthScore: number;
}