# lib/api-zod/src/generated/types/environmentMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */

export interface EnvironmentMetrics {
  temperature: number;
  humidity: number;
  vibration: number;
  power: number;
  timestamp: string;
}