# lib/api-zod/src/generated/types/equipment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */
import type { EquipmentStatus } from './equipmentStatus';
import type { Metric } from './metric';

export interface Equipment {
  id: string;
  /** Chinese name */
  name: string;
  /** English name */
  nameEn: string;
  category: string;
  status: EquipmentStatus;
  location: string;
  locationEn?: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  lastMaintenance: string;
  /** @nullable */
  predictedFailureDate?: string | null;
  metrics?: Metric[];
}