# lib/api-zod/src/generated/types/equipmentCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentCategory {
  id: string;
  name: string;
  nameEn: string;
  icon: string;
  count: number;
  alertCount: number;
}