# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */
import type { MetricTrend } from './metricTrend';

export interface Metric {
  key: string;
  label: string;
  value: number;
  unit: string;
  timestamp: string;
  /** @nullable */
  trend?: MetricTrend;
}