# lib/api-zod/src/generated/types/statusBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Equipment Management Digital Twin API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusBreakdown {
  running: number;
  warning: number;
  fault: number;
  offline: number;
  maintenance: number;
}